Please read the entire commit message:
This commit refactors all of the classes that previously inherited from rectangle to now inherit from actor. The gamescreen class has been remade to account for this, however a lot of the former functionality needs reimplementing.

The 'old' game screen class is available in /core.

Train classes will also need to be rewritten to function as an actor and not as a rectangle.

The HUD can be implemented as its own stage.